refactor(server): extract response helpers into response.rs
