+++ crates/server/src/pipeline_handlers.rs
@@ -11,7 +11,6 @@
 //! * `POST /:name/execute`  — run a pipeline's SELECT with bound parameters
 
 use anyhow::Result;
-use arrow::record_batch::RecordBatch;
 use axum::{
     Json,
     extract::{Path, State},
@@ -26,6 +25,9 @@ use std::collections::HashMap;
 use std::time::Instant;
 
 use crate::config::DataSourceType;
+use crate::response::{
+    ErrorResponse, create_error_response, create_success_response, record_batch_to_json,
+};
 use crate::semantics::SemanticsRegistry;
 use crate::server::AppState;
 
@@ -46,21 +48,6 @@ pub struct ExecuteResponse {
     pub rows: usize,
     /// Execution time in milliseconds
     pub execution_time_ms: u64,
-}
-
-/// Error response structure for API endpoints
-#[derive(Debug, Serialize)]
-pub struct ErrorResponse {
-    /// Whether the operation was successful
-    pub success: bool,
-    /// Error message
-    pub error: String,
-    /// Error category/type
-    pub error_type: String,
-    /// Additional error details
-    pub details: Option<Value>,
-    /// Timestamp when error occurred
-    pub timestamp: String,
 }
 
 /// Field information for table schema
@@ -104,32 +91,6 @@ pub struct DataSourceResponse {
     pub url: Option<String>,
     /// Registered tables with their schemas
     pub tables: Vec<TableInfo>,
-}
-
-/// Helper function to create error responses
-fn create_error_response(
-    error_msg: &str,
-    error_type: &str,
-    details: Option<Value>,
-) -> Json<ErrorResponse> {
-    Json(ErrorResponse {
-        success: false,
-        error: error_msg.to_string(),
-        error_type: error_type.to_string(),
-        details,
-        timestamp: chrono::Utc::now().to_rfc3339(),
-    })
-}
-
-/// Helper function to create success response with data
-fn create_success_response(data: Vec<Value>, rows: usize, execution_time_ms: u64) -> Json<Value> {
-    Json(serde_json::json!({
-        "success": true,
-        "data": data,
-        "rows": rows,
-        "execution_time_ms": execution_time_ms,
-        "timestamp": chrono::Utc::now().to_rfc3339()
-    }))
 }
 
 /// Get table schema from SessionContext
@@ -824,33 +785,12 @@ pub async fn execute_pipeline_by_name(
         execution_time
     );
 
-    Ok(create_success_response(data, row_count, execution_time))
-}
-
-/// Convert Arrow RecordBatch to JSON array using arrow_json
-fn record_batch_to_json(batch: &RecordBatch) -> Result<Vec<Value>, Box<dyn std::error::Error>> {
-    use arrow_json::{WriterBuilder, writer::JsonArray};
-    use serde_json::Map;
-
-    // Write the record batch to JSON using arrow_json with null value inclusion
-    let buf = Vec::new();
-    let mut writer = WriterBuilder::new()
-        .with_explicit_nulls(true) // Include null values in JSON output
-        .build::<_, JsonArray>(buf);
-    writer.write_batches(&vec![batch])?;
-    writer.finish()?;
-    let json_data = writer.into_inner();
-
-    // Parse the JSON array string into serde_json::Value objects
-    let json_rows: Vec<Map<String, Value>> = serde_json::from_reader(json_data.as_slice())?;
-
-    // Convert Map objects to Value objects
-    let values: Vec<Value> = json_rows
-        .into_iter()
-        .map(|map| Value::Object(map))
-        .collect();
-
-    Ok(values)
+    Ok(create_success_response(
+        data,
+        row_count,
+        execution_time,
+        None,
+    ))
 }
 
 #[cfg(test)]
